refactor: organize e2e tests by bounded context
Reorganize E2E tests into context-based directories to align with src/contexts and improve test ownership clarity.

Made-with: Cursor
