+++ tests/e2e/configuration/mod.rs
@@ -1,0 +1,1 @@
+mod config;
