+++ tests/e2e/interface/mod.rs
@@ -1,0 +1,1 @@
+mod cli;
