+++ tests/e2e/merge_readiness/merge_ready.rs
@@ -3,7 +3,7 @@
 //! `✓ merge-ready` が表示される条件（全ブロッカーが存在しない）と、
 //! 全ブロッカーが同時成立したときに `merge-ready` が表示されないことを検証する。
 
-use super::helpers::TestEnv;
+use super::super::helpers::TestEnv;
 use assert_cmd::Command;
 
 fn cmd(env: &TestEnv) -> Command {
