+++ tests/e2e/merge_readiness/mod.rs
@@ -1,0 +1,6 @@
+mod branch_sync;
+mod ci_checks;
+mod errors;
+mod merge_ready;
+mod pr_state;
+mod review;
