+++ tests/e2e/status_cache/mod.rs
@@ -1,0 +1,2 @@
+mod cache;
+mod daemon;
